Added statistics tracking for # of Questions, # correct answers, # incorrect answers, and response time
Added statistics tracking for # of Questions, # correct answers, # incorrect answers, and response time. Using a SQLlite db handled directly with python. Right now a new .db file for the stats is created every time, similar to how the embedding work on main branch. Can change it to load a .db file in the future so statistics are persistent.